feat: update iron-session version to 8 (#14)

+++ src/pages/api/category/remove.ts
@@ -1,5 +1,7 @@
-import type { NextApiHandler } from "next";
-import { withSessionRoute } from "y/server/wrap";
+import {
+  type NextApiHandlerWithSession,
+  withSessionRoute,
+} from "y/server/wrap";
 import { prisma } from "y/server/db";
 import { createSuccessRes } from "y/utils/apiResponse";
 
@@ -7,9 +9,10 @@ export interface RemoveCategoryParam {
   id: number;
 }
 
-const handler: NextApiHandler = async (req, res) => {
+const handler: NextApiHandlerWithSession = async (req, res, session) => {
+  const userId = session.user.id;
   const data = req.body as RemoveCategoryParam;
-  await prisma.category.delete({ where: { id: data.id } });
+  await prisma.category.delete({ where: { id: data.id, userId } });
   return createSuccessRes(res, null);
 };
